% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{Annotation.info}
\alias{Annotation.info}
\title{Annotation.info Object}
\usage{
Annotation.info()
}
\value{
Annotation.info object
}
\description{
Annotation.info Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A map of additional read alignment information. This must be of the form map (string key mapping to a list of string values).
}
\seealso{
Other Annotation functions: \code{\link{Annotation}},
  \code{\link{annotations.create}},
  \code{\link{annotations.update}}
}

