% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{Annotation}
\alias{Annotation}
\title{Annotation Object}
\usage{
Annotation(Annotation.info = NULL, transcript = NULL, start = NULL,
  annotationSetId = NULL, name = NULL, variant = NULL,
  referenceId = NULL, id = NULL, reverseStrand = NULL,
  referenceName = NULL, info = NULL, type = NULL, end = NULL)
}
\arguments{
\item{Annotation.info}{The \link{Annotation.info} object or list of objects}

\item{transcript}{A transcript value represents the assertion that a particular region of}

\item{start}{The start position of the range on the reference, 0-based inclusive}

\item{annotationSetId}{The annotation set to which this annotation belongs}

\item{name}{The display name of this annotation}

\item{variant}{A variant annotation, which describes the effect of a variant on the}

\item{referenceId}{The ID of the Google Genomics reference associated with this range}

\item{id}{The server-generated annotation ID, unique across all annotations}

\item{reverseStrand}{Whether this range refers to the reverse strand, as opposed to the forward}

\item{referenceName}{The display name corresponding to the reference specified by}

\item{info}{A map of additional read alignment information}

\item{type}{The data type for this annotation}

\item{end}{The end position of the range on the reference, 0-based exclusive}
}
\value{
Annotation object
}
\description{
Annotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An annotation describes a region of reference genome. The value of anannotation may be one of several canonical types, supplemented by arbitraryinfo tags. An annotation is not inherently associated with a specificsample or individual (though a client could choose to use annotations inthis way). Example canonical annotation types are `GENE` and`VARIANT`.
}
\seealso{
Other Annotation functions: \code{\link{Annotation.info}},
  \code{\link{annotations.create}},
  \code{\link{annotations.update}}
}
