% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_functions.R
\name{annotations.batchCreate}
\alias{annotations.batchCreate}
\title{Creates one or more new annotations atomically. All annotations mustbelong to the same annotation set. Caller must have WRITEpermission for this annotation set. For optimal performance, batchpositionally adjacent annotations together.If the request has a systemic issue, such as an attempt to write toan inaccessible annotation set, the entire RPC will fail accordingly. Forlesser data issues, when possible an error will be isolated to thecorresponding batch entry in the response; the remaining well formedannotations will be created normally.For details on the requirements for each individual annotation resource,seeCreateAnnotation.}
\usage{
annotations.batchCreate(BatchCreateAnnotationsRequest)
}
\arguments{
\item{BatchCreateAnnotationsRequest}{The \link{BatchCreateAnnotationsRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/genomics
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/genomics)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/genomics}{Google Documentation}

Other BatchCreateAnnotationsRequest functions: \code{\link{BatchCreateAnnotationsRequest}}
}
