% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_functions.R
\name{annotations.search}
\alias{annotations.search}
\title{Searches for annotations that match the given criteria. Results areordered by genomic coordinate (by reference sequence, then position).Annotations with equivalent genomic coordinates are returned in anunspecified order. This order is consistent, such that two queries for thesame content (regardless of page size) yield annotations in the same orderacross their respective streams of paginated responses. Caller must haveREAD permission for the queried annotation sets.}
\usage{
annotations.search(SearchAnnotationsRequest)
}
\arguments{
\item{SearchAnnotationsRequest}{The \link{SearchAnnotationsRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/genomics
\item https://www.googleapis.com/auth/genomics.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/genomics, https://www.googleapis.com/auth/genomics.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/genomics}{Google Documentation}

Other SearchAnnotationsRequest functions: \code{\link{SearchAnnotationsRequest}}
}
