% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_functions.R
\name{annotations.update}
\alias{annotations.update}
\title{Updates an annotation. Caller must haveWRITE permission for the associated dataset.}
\usage{
annotations.update(Annotation, annotationId, updateMask = NULL)
}
\arguments{
\item{Annotation}{The \link{Annotation} object to pass to this method}

\item{annotationId}{The ID of the annotation to be updated}

\item{updateMask}{An optional mask specifying which fields to update}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/genomics
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/genomics)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/genomics}{Google Documentation}

Other Annotation functions: \code{\link{Annotation.info}},
  \code{\link{Annotation}},
  \code{\link{annotations.create}}
}
