% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_functions.R
\name{annotationsets.create}
\alias{annotationsets.create}
\title{Creates a new annotation set. Caller must have WRITE permission for theassociated dataset.The following fields are required:  * datasetId  * referenceSetIdAll other fields may be optionally specified, unless documented as beingserver-generated (for example, the `id` field).}
\usage{
annotationsets.create(AnnotationSet)
}
\arguments{
\item{AnnotationSet}{The \link{AnnotationSet} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/genomics
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/genomics)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/genomics}{Google Documentation}

Other AnnotationSet functions: \code{\link{AnnotationSet.info}},
  \code{\link{AnnotationSet}},
  \code{\link{annotationsets.update}}
}
