% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_functions.R
\name{annotationsets.update}
\alias{annotationsets.update}
\title{Updates an annotation set. The update must respect all mutability restrictions and other invariants described on the annotation set resource. Caller must have WRITE permission for the associated dataset.}
\usage{
annotationsets.update(AnnotationSet, annotationSetId, updateMask = NULL)
}
\arguments{
\item{AnnotationSet}{The \link{AnnotationSet} object to pass to this method}

\item{annotationSetId}{The ID of the annotation set to be updated}

\item{updateMask}{An optional mask specifying which fields to update}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/genomics
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/genomics)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/genomics/}{Google Documentation}

Other AnnotationSet functions: \code{\link{AnnotationSet.info}},
  \code{\link{AnnotationSet}},
  \code{\link{annotationsets.create}}
}

