% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{BatchCreateAnnotationsRequest}
\alias{BatchCreateAnnotationsRequest}
\title{BatchCreateAnnotationsRequest Object}
\usage{
BatchCreateAnnotationsRequest(annotations = NULL, requestId = NULL)
}
\arguments{
\item{annotations}{The annotations to be created}

\item{requestId}{A unique request ID which enables the server to detect duplicated requests}
}
\value{
BatchCreateAnnotationsRequest object
}
\description{
BatchCreateAnnotationsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other BatchCreateAnnotationsRequest functions: \code{\link{annotations.batchCreate}}
}

