% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{BatchCreateAnnotationsResponse}
\alias{BatchCreateAnnotationsResponse}
\title{BatchCreateAnnotationsResponse Object}
\usage{
BatchCreateAnnotationsResponse(entries = NULL)
}
\arguments{
\item{entries}{The resulting per-annotation entries, ordered consistently with the original request}
}
\value{
BatchCreateAnnotationsResponse object
}
\description{
BatchCreateAnnotationsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

