% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{CallSet.info}
\alias{CallSet.info}
\title{CallSet.info Object}
\usage{
CallSet.info()
}
\value{
CallSet.info object
}
\description{
CallSet.info Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A map of additional call set information. This must be of the form map (string key mapping to a list of string values).
}
\seealso{
Other CallSet functions: \code{\link{CallSet}},
  \code{\link{callsets.create}},
  \code{\link{callsets.patch}}
}

