% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{CallSet}
\alias{CallSet}
\title{CallSet Object}
\usage{
CallSet(CallSet.info = NULL, id = NULL, name = NULL, sampleId = NULL,
  variantSetIds = NULL, created = NULL, info = NULL)
}
\arguments{
\item{CallSet.info}{The \link{CallSet.info} object or list of objects}

\item{id}{The server-generated call set ID, unique across all call sets}

\item{name}{The call set name}

\item{sampleId}{The sample ID this call set corresponds to}

\item{variantSetIds}{The IDs of the variant sets this call set belongs to}

\item{created}{The date this call set was created in milliseconds from the epoch}

\item{info}{A map of additional call set information}
}
\value{
CallSet object
}
\description{
CallSet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A call set is a collection of variant calls, typically for one sample. It belongs to a variant set. For more genomics resource definitions, see [Fundamentals of Google Genomics](https://cloud.google.com/genomics/fundamentals-of-google-genomics)
}
\seealso{
Other CallSet functions: \code{\link{CallSet.info}},
  \code{\link{callsets.create}},
  \code{\link{callsets.patch}}
}

