% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_functions.R
\name{callsets.patch}
\alias{callsets.patch}
\title{Updates a call set.For the definitions of call sets and other genomics resources, see[Fundamentals of GoogleGenomics](https://cloud.google.com/genomics/fundamentals-of-google-genomics)This method supports patch semantics.}
\usage{
callsets.patch(CallSet, callSetId, updateMask = NULL)
}
\arguments{
\item{CallSet}{The \link{CallSet} object to pass to this method}

\item{callSetId}{The ID of the call set to be updated}

\item{updateMask}{An optional mask specifying which fields to update}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/genomics
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/genomics)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/genomics}{Google Documentation}

Other CallSet functions: \code{\link{CallSet.info}},
  \code{\link{CallSet}}, \code{\link{callsets.create}}
}
