% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_functions.R
\name{callsets.search}
\alias{callsets.search}
\title{Gets a list of call sets matching the criteria.For the definitions of call sets and other genomics resources, see[Fundamentals of GoogleGenomics](https://cloud.google.com/genomics/fundamentals-of-google-genomics)Implements[GlobalAllianceApi.searchCallSets](https://github.com/ga4gh/schemas/blob/v0.5.1/src/main/resources/avro/variantmethods.avdl#L178).}
\usage{
callsets.search(SearchCallSetsRequest)
}
\arguments{
\item{SearchCallSetsRequest}{The \link{SearchCallSetsRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/genomics
\item https://www.googleapis.com/auth/genomics.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/genomics, https://www.googleapis.com/auth/genomics.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/genomics}{Google Documentation}

Other SearchCallSetsRequest functions: \code{\link{SearchCallSetsRequest}}
}
