% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{CodingSequence}
\alias{CodingSequence}
\title{CodingSequence Object}
\usage{
CodingSequence(start = NULL, end = NULL)
}
\arguments{
\item{start}{The start of the coding sequence on this annotation's reference sequence,}

\item{end}{The end of the coding sequence on this annotation's reference sequence,}
}
\value{
CodingSequence object
}
\description{
CodingSequence Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
