% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{CoverageBucket}
\alias{CoverageBucket}
\title{CoverageBucket Object}
\usage{
CoverageBucket(meanCoverage = NULL, range = NULL)
}
\arguments{
\item{meanCoverage}{The average number of reads which are aligned to each individual}

\item{range}{The genomic coordinate range spanned by this bucket}
}
\value{
CoverageBucket object
}
\description{
CoverageBucket Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A bucket over which read coverage has been precomputed. A bucket correspondsto a specific range of the reference sequence.
}
