% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_functions.R
\name{datasets.getIamPolicy}
\alias{datasets.getIamPolicy}
\title{Gets the access control policy for the dataset. This is empty if thepolicy or resource does not exist.See <a href='/iam/docs/managing-policies#getting_a_policy'>Getting aPolicy</a> for more information.For the definitions of datasets and other genomics resources, see[Fundamentals of GoogleGenomics](https://cloud.google.com/genomics/fundamentals-of-google-genomics)}
\usage{
datasets.getIamPolicy(GetIamPolicyRequest, resource)
}
\arguments{
\item{GetIamPolicyRequest}{The \link{GetIamPolicyRequest} object to pass to this method}

\item{resource}{REQUIRED: The resource for which policy is being specified}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/genomics
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/genomics)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/genomics}{Google Documentation}

Other GetIamPolicyRequest functions: \code{\link{GetIamPolicyRequest}}
}
