% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{Entry}
\alias{Entry}
\title{Entry Object}
\usage{
Entry(status = NULL, annotation = NULL)
}
\arguments{
\item{status}{The creation status}

\item{annotation}{The created annotation, if creation was successful}
}
\value{
Entry object
}
\description{
Entry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

