% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{Exon}
\alias{Exon}
\title{Exon Object}
\usage{
Exon(start = NULL, end = NULL, frame = NULL)
}
\arguments{
\item{start}{The start position of the exon on this annotation's reference sequence, 0-based inclusive}

\item{end}{The end position of the exon on this annotation's reference sequence, 0-based exclusive}

\item{frame}{The frame of this exon}
}
\value{
Exon object
}
\description{
Exon Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

