% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{Experiment}
\alias{Experiment}
\title{Experiment Object}
\usage{
Experiment(libraryId = NULL, platformUnit = NULL, sequencingCenter = NULL,
  instrumentModel = NULL)
}
\arguments{
\item{libraryId}{A client-supplied library identifier; a library is a collection of DNA fragments which have been prepared for sequencing from a sample}

\item{platformUnit}{The platform unit used as part of this experiment, for example flowcell-barcode}

\item{sequencingCenter}{The sequencing center used as part of this experiment}

\item{instrumentModel}{The instrument model used as part of this experiment}
}
\value{
Experiment object
}
\description{
Experiment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

