% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ExportVariantSetRequest}
\alias{ExportVariantSetRequest}
\title{Genomics API Objects 
Upload, process, query, and search Genomics data in the cloud.}
\usage{
ExportVariantSetRequest(callSetIds = NULL, projectId = NULL,
  format = NULL, bigqueryDataset = NULL, bigqueryTable = NULL)
}
\arguments{
\item{callSetIds}{If provided, only variant call information from the specified call sets}

\item{projectId}{Required}

\item{format}{The format for the exported data}

\item{bigqueryDataset}{Required}

\item{bigqueryTable}{Required}
}
\value{
ExportVariantSetRequest object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:53:39
filename: /Users/mark/dev/R/autoGoogleAPI/googlegenomicsv1.auto/R/genomics_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
ExportVariantSetRequest Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The variant data export request.
}
\seealso{
Other ExportVariantSetRequest functions: \code{\link{variantsets.export}}
}
