% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ImportReadGroupSetsRequest}
\alias{ImportReadGroupSetsRequest}
\title{ImportReadGroupSetsRequest Object}
\usage{
ImportReadGroupSetsRequest(datasetId = NULL, referenceSetId = NULL,
  sourceUris = NULL, partitionStrategy = NULL)
}
\arguments{
\item{datasetId}{Required}

\item{referenceSetId}{The reference set to which the imported read group sets are aligned to, if any}

\item{sourceUris}{A list of URIs pointing at [BAM files](https://samtools}

\item{partitionStrategy}{The partition strategy describes how read groups are partitioned into read group sets}
}
\value{
ImportReadGroupSetsRequest object
}
\description{
ImportReadGroupSetsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The read group set import request.
}
\seealso{
Other ImportReadGroupSetsRequest functions: \code{\link{readgroupsets.import}}
}

