% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ImportVariantsRequest.infoMergeConfig}
\alias{ImportVariantsRequest.infoMergeConfig}
\title{ImportVariantsRequest.infoMergeConfig Object}
\usage{
ImportVariantsRequest.infoMergeConfig()
}
\value{
ImportVariantsRequest.infoMergeConfig object
}
\description{
ImportVariantsRequest.infoMergeConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A mapping between info field keys and the InfoMergeOperations tobe performed on them. This is plumbed down to the MergeVariantRequestsgenerated by the resulting import job.
}
\seealso{
Other ImportVariantsRequest functions: \code{\link{ImportVariantsRequest}},
  \code{\link{variants.import}}
}
