% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ImportVariantsResponse}
\alias{ImportVariantsResponse}
\title{ImportVariantsResponse Object}
\usage{
ImportVariantsResponse(callSetIds = NULL)
}
\arguments{
\item{callSetIds}{IDs of the call sets created during the import}
}
\value{
ImportVariantsResponse object
}
\description{
ImportVariantsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The variant data import response.
}
