% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{LinearAlignment}
\alias{LinearAlignment}
\title{LinearAlignment Object}
\usage{
LinearAlignment(mappingQuality = NULL, position = NULL, cigar = NULL)
}
\arguments{
\item{mappingQuality}{The mapping quality of this alignment}

\item{position}{The position of this alignment}

\item{cigar}{Represents the local alignment of this sequence (alignment matches, indels,}
}
\value{
LinearAlignment object
}
\description{
LinearAlignment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A linear alignment can be represented by one CIGAR string. Describes themapped position and local alignment of the read to the reference.
}
