% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ListBasesResponse}
\alias{ListBasesResponse}
\title{ListBasesResponse Object}
\usage{
ListBasesResponse(sequence = NULL, offset = NULL, nextPageToken = NULL)
}
\arguments{
\item{sequence}{A substring of the bases that make up this reference}

\item{offset}{The offset position (0-based) of the given `sequence` from the}

\item{nextPageToken}{The continuation token, which is used to page through large result sets}
}
\value{
ListBasesResponse object
}
\description{
ListBasesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
