% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ListDatasetsResponse}
\alias{ListDatasetsResponse}
\title{ListDatasetsResponse Object}
\usage{
ListDatasetsResponse(datasets = NULL, nextPageToken = NULL)
}
\arguments{
\item{datasets}{The list of matching Datasets}

\item{nextPageToken}{The continuation token, which is used to page through large result sets}
}
\value{
ListDatasetsResponse object
}
\description{
ListDatasetsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The dataset list response.
}
