% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{MergeVariantsRequest.infoMergeConfig}
\alias{MergeVariantsRequest.infoMergeConfig}
\title{MergeVariantsRequest.infoMergeConfig Object}
\usage{
MergeVariantsRequest.infoMergeConfig()
}
\value{
MergeVariantsRequest.infoMergeConfig object
}
\description{
MergeVariantsRequest.infoMergeConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A mapping between info field keys and the InfoMergeOperations tobe performed on them.
}
\seealso{
Other MergeVariantsRequest functions: \code{\link{MergeVariantsRequest}},
  \code{\link{variants.merge}}
}
