% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{MergeVariantsRequest}
\alias{MergeVariantsRequest}
\title{MergeVariantsRequest Object}
\usage{
MergeVariantsRequest(MergeVariantsRequest.infoMergeConfig = NULL,
  variantSetId = NULL, variants = NULL, infoMergeConfig = NULL)
}
\arguments{
\item{MergeVariantsRequest.infoMergeConfig}{The \link{MergeVariantsRequest.infoMergeConfig} object or list of objects}

\item{variantSetId}{The destination variant set}

\item{variants}{The variants to be merged with existing variants}

\item{infoMergeConfig}{A mapping between info field keys and the InfoMergeOperations to be performed on them}
}
\value{
MergeVariantsRequest object
}
\description{
MergeVariantsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other MergeVariantsRequest functions: \code{\link{MergeVariantsRequest.infoMergeConfig}},
  \code{\link{variants.merge}}
}

