% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{OperationMetadata.labels}
\alias{OperationMetadata.labels}
\title{OperationMetadata.labels Object}
\usage{
OperationMetadata.labels()
}
\value{
OperationMetadata.labels object
}
\description{
OperationMetadata.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Optionally provided by the caller when submitting the request that createsthe operation.
}
\seealso{
Other OperationMetadata functions: \code{\link{OperationMetadata.request}},
  \code{\link{OperationMetadata.runtimeMetadata}},
  \code{\link{OperationMetadata}}
}
