% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{OperationMetadata.runtimeMetadata}
\alias{OperationMetadata.runtimeMetadata}
\title{OperationMetadata.runtimeMetadata Object}
\usage{
OperationMetadata.runtimeMetadata()
}
\value{
OperationMetadata.runtimeMetadata object
}
\description{
OperationMetadata.runtimeMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Runtime metadata on this Operation.
}
\seealso{
Other OperationMetadata functions: \code{\link{OperationMetadata.request}},
  \code{\link{OperationMetadata}}
}

