% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{Program}
\alias{Program}
\title{Program Object}
\usage{
Program(commandLine = NULL, prevProgramId = NULL, id = NULL,
  version = NULL, name = NULL)
}
\arguments{
\item{commandLine}{The command line used to run this program}

\item{prevProgramId}{The ID of the program run before this one}

\item{id}{The user specified locally unique ID of the program}

\item{version}{The version of the program run}

\item{name}{The display name of the program}
}
\value{
Program object
}
\description{
Program Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
