% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{Read.info}
\alias{Read.info}
\title{Read.info Object}
\usage{
Read.info()
}
\value{
Read.info object
}
\description{
Read.info Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A map of additional read alignment information. This must be of the formmap<string, string[]> (string key mapping to a list of string values).
}
\seealso{
Other Read functions: \code{\link{Read}}
}
