% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ReadGroup.info}
\alias{ReadGroup.info}
\title{ReadGroup.info Object}
\usage{
ReadGroup.info()
}
\value{
ReadGroup.info object
}
\description{
ReadGroup.info Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A map of additional read group information. This must be of the formmap<string, string[]> (string key mapping to a list of string values).
}
\seealso{
Other ReadGroup functions: \code{\link{ReadGroup}}
}
