% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ReadGroupSet.info}
\alias{ReadGroupSet.info}
\title{ReadGroupSet.info Object}
\usage{
ReadGroupSet.info()
}
\value{
ReadGroupSet.info object
}
\description{
ReadGroupSet.info Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A map of additional read group set information.
}
\seealso{
Other ReadGroupSet functions: \code{\link{ReadGroupSet}},
  \code{\link{readgroupsets.patch}}
}

