% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ReadGroupSet}
\alias{ReadGroupSet}
\title{ReadGroupSet Object}
\usage{
ReadGroupSet(ReadGroupSet.info = NULL, id = NULL, datasetId = NULL,
  referenceSetId = NULL, name = NULL, filename = NULL,
  readGroups = NULL, info = NULL)
}
\arguments{
\item{ReadGroupSet.info}{The \link{ReadGroupSet.info} object or list of objects}

\item{id}{The server-generated read group set ID, unique for all read group sets}

\item{datasetId}{The dataset to which this read group set belongs}

\item{referenceSetId}{The reference set to which the reads in this read group set are aligned}

\item{name}{The read group set name}

\item{filename}{The filename of the original source file for this read group set, if any}

\item{readGroups}{The read groups in this set}

\item{info}{A map of additional read group set information}
}
\value{
ReadGroupSet object
}
\description{
ReadGroupSet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A read group set is a logical collection of read groups, which are collections of reads produced by a sequencer. A read group set typically models reads corresponding to one sample, sequenced one way, and aligned one way. * A read group set belongs to one dataset. * A read group belongs to one read group set. * A read belongs to one read group. For more genomics resource definitions, see [Fundamentals of Google Genomics](https://cloud.google.com/genomics/fundamentals-of-google-genomics)
}
\seealso{
Other ReadGroupSet functions: \code{\link{ReadGroupSet.info}},
  \code{\link{readgroupsets.patch}}
}

