% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_functions.R
\name{readgroupsets.export}
\alias{readgroupsets.export}
\title{Exports a read group set to a BAM file in Google Cloud Storage. For the definitions of read group sets and other genomics resources, see [Fundamentals of Google Genomics](https://cloud.google.com/genomics/fundamentals-of-google-genomics) Note that currently there may be some differences between exported BAM files and the original BAM file at the time of import. See ImportReadGroupSets for caveats.}
\usage{
readgroupsets.export(ExportReadGroupSetRequest, readGroupSetId)
}
\arguments{
\item{ExportReadGroupSetRequest}{The \link{ExportReadGroupSetRequest} object to pass to this method}

\item{readGroupSetId}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/devstorage.read_write
\item https://www.googleapis.com/auth/genomics
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/devstorage.read_write, https://www.googleapis.com/auth/genomics)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/genomics/}{Google Documentation}

Other ExportReadGroupSetRequest functions: \code{\link{ExportReadGroupSetRequest}}
}

