% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_functions.R
\name{readgroupsets.patch}
\alias{readgroupsets.patch}
\title{Updates a read group set.For the definitions of read group sets and other genomics resources, see[Fundamentals of GoogleGenomics](https://cloud.google.com/genomics/fundamentals-of-google-genomics)This method supports patch semantics.}
\usage{
readgroupsets.patch(ReadGroupSet, readGroupSetId, updateMask = NULL)
}
\arguments{
\item{ReadGroupSet}{The \link{ReadGroupSet} object to pass to this method}

\item{readGroupSetId}{The ID of the read group set to be updated}

\item{updateMask}{An optional mask specifying which fields to update}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/genomics
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/genomics)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/genomics}{Google Documentation}

Other ReadGroupSet functions: \code{\link{ReadGroupSet.info}},
  \code{\link{ReadGroupSet}}
}
