% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_functions.R
\name{reads.search}
\alias{reads.search}
\title{Gets a list of reads for one or more read group sets.For the definitions of read group sets and other genomics resources, see[Fundamentals of GoogleGenomics](https://cloud.google.com/genomics/fundamentals-of-google-genomics)Reads search operates over a genomic coordinate space of reference sequence& position defined over the reference sequences to which the requestedread group sets are aligned.If a target positional range is specified, search returns all reads whosealignment to the reference genome overlap the range. A query whichspecifies only read group set IDs yields all reads in those read groupsets, including unmapped reads.All reads returned (including reads on subsequent pages) are ordered bygenomic coordinate (by reference sequence, then position). Reads withequivalent genomic coordinates are returned in an unspecified order. Thisorder is consistent, such that two queries for the same content (regardlessof page size) yield reads in the same order across their respective streamsof paginated responses.Implements[GlobalAllianceApi.searchReads](https://github.com/ga4gh/schemas/blob/v0.5.1/src/main/resources/avro/readmethods.avdl#L85).}
\usage{
reads.search(SearchReadsRequest)
}
\arguments{
\item{SearchReadsRequest}{The \link{SearchReadsRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/genomics
\item https://www.googleapis.com/auth/genomics.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/genomics, https://www.googleapis.com/auth/genomics.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/genomics}{Google Documentation}

Other SearchReadsRequest functions: \code{\link{SearchReadsRequest}}
}
