% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{Reference}
\alias{Reference}
\title{Reference Object}
\usage{
Reference(ncbiTaxonId = NULL, sourceUri = NULL, name = NULL,
  md5checksum = NULL, id = NULL, length = NULL, sourceAccessions = NULL)
}
\arguments{
\item{ncbiTaxonId}{ID from http://www}

\item{sourceUri}{The URI from which the sequence was obtained}

\item{name}{The name of this reference, for example `22`}

\item{md5checksum}{MD5 of the upper-case sequence excluding all whitespace characters (this}

\item{id}{The server-generated reference ID, unique across all references}

\item{length}{The length of this reference's sequence}

\item{sourceAccessions}{All known corresponding accession IDs in INSDC (GenBank/ENA/DDBJ) ideally}
}
\value{
Reference object
}
\description{
Reference Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A reference is a canonical assembled DNA sequence, intended to act as areference coordinate space for other genomic annotations. A single referencemight represent the human chromosome 1 or mitochandrial DNA, for instance. Areference belongs to one or more reference sets.For more genomics resource definitions, see [Fundamentals of GoogleGenomics](https://cloud.google.com/genomics/fundamentals-of-google-genomics)
}
