% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ReferenceBound}
\alias{ReferenceBound}
\title{ReferenceBound Object}
\usage{
ReferenceBound(upperBound = NULL, referenceName = NULL)
}
\arguments{
\item{upperBound}{An upper bound (inclusive) on the starting coordinate of any}

\item{referenceName}{The name of the reference associated with this reference bound}
}
\value{
ReferenceBound object
}
\description{
ReferenceBound Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
ReferenceBound records an upper bound for the starting coordinate ofvariants in a particular reference.
}
