% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{RuntimeMetadata}
\alias{RuntimeMetadata}
\title{RuntimeMetadata Object}
\usage{
RuntimeMetadata(computeEngine = NULL)
}
\arguments{
\item{computeEngine}{Execution information specific to Google Compute Engine}
}
\value{
RuntimeMetadata object
}
\description{
RuntimeMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Runtime metadata that will be populated in theruntimeMetadatafield of the Operation associated with a RunPipeline execution.
}
