% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{SearchAnnotationSetsResponse}
\alias{SearchAnnotationSetsResponse}
\title{SearchAnnotationSetsResponse Object}
\usage{
SearchAnnotationSetsResponse(nextPageToken = NULL, annotationSets = NULL)
}
\arguments{
\item{nextPageToken}{The continuation token, which is used to page through large result sets}

\item{annotationSets}{The matching annotation sets}
}
\value{
SearchAnnotationSetsResponse object
}
\description{
SearchAnnotationSetsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
