% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{SearchAnnotationsResponse}
\alias{SearchAnnotationsResponse}
\title{SearchAnnotationsResponse Object}
\usage{
SearchAnnotationsResponse(nextPageToken = NULL, annotations = NULL)
}
\arguments{
\item{nextPageToken}{The continuation token, which is used to page through large result sets}

\item{annotations}{The matching annotations}
}
\value{
SearchAnnotationsResponse object
}
\description{
SearchAnnotationsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
