% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{SearchCallSetsResponse}
\alias{SearchCallSetsResponse}
\title{SearchCallSetsResponse Object}
\usage{
SearchCallSetsResponse(nextPageToken = NULL, callSets = NULL)
}
\arguments{
\item{nextPageToken}{The continuation token, which is used to page through large result sets}

\item{callSets}{The list of matching call sets}
}
\value{
SearchCallSetsResponse object
}
\description{
SearchCallSetsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The call set search response.
}
