% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{SearchReadGroupSetsResponse}
\alias{SearchReadGroupSetsResponse}
\title{SearchReadGroupSetsResponse Object}
\usage{
SearchReadGroupSetsResponse(readGroupSets = NULL, nextPageToken = NULL)
}
\arguments{
\item{readGroupSets}{The list of matching read group sets}

\item{nextPageToken}{The continuation token, which is used to page through large result sets}
}
\value{
SearchReadGroupSetsResponse object
}
\description{
SearchReadGroupSetsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The read group set search response.
}

