% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{SearchReferenceSetsResponse}
\alias{SearchReferenceSetsResponse}
\title{SearchReferenceSetsResponse Object}
\usage{
SearchReferenceSetsResponse(referenceSets = NULL, nextPageToken = NULL)
}
\arguments{
\item{referenceSets}{The matching references sets}

\item{nextPageToken}{The continuation token, which is used to page through large result sets}
}
\value{
SearchReferenceSetsResponse object
}
\description{
SearchReferenceSetsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

