% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{SearchReferencesRequest}
\alias{SearchReferencesRequest}
\title{SearchReferencesRequest Object}
\usage{
SearchReferencesRequest(md5checksums = NULL, accessions = NULL,
  referenceSetId = NULL, pageToken = NULL, pageSize = NULL)
}
\arguments{
\item{md5checksums}{If present, return references for which the md5checksum matches exactly}

\item{accessions}{If present, return references for which a prefix of any of sourceAccessions match any of these strings}

\item{referenceSetId}{If present, return only references which belong to this reference set}

\item{pageToken}{The continuation token, which is used to page through large result sets}

\item{pageSize}{The maximum number of results to return in a single page}
}
\value{
SearchReferencesRequest object
}
\description{
SearchReferencesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other SearchReferencesRequest functions: \code{\link{references.search}}
}

