% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{SearchReferencesResponse}
\alias{SearchReferencesResponse}
\title{SearchReferencesResponse Object}
\usage{
SearchReferencesResponse(references = NULL, nextPageToken = NULL)
}
\arguments{
\item{references}{The matching references}

\item{nextPageToken}{The continuation token, which is used to page through large result sets}
}
\value{
SearchReferencesResponse object
}
\description{
SearchReferencesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

