% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{SearchVariantSetsResponse}
\alias{SearchVariantSetsResponse}
\title{SearchVariantSetsResponse Object}
\usage{
SearchVariantSetsResponse(nextPageToken = NULL, variantSets = NULL)
}
\arguments{
\item{nextPageToken}{The continuation token, which is used to page through large result sets}

\item{variantSets}{The variant sets belonging to the requested dataset}
}
\value{
SearchVariantSetsResponse object
}
\description{
SearchVariantSetsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The search variant sets response.
}
