% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{StreamReadsRequest}
\alias{StreamReadsRequest}
\title{StreamReadsRequest Object}
\usage{
StreamReadsRequest(projectId = NULL, readGroupSetId = NULL,
  referenceName = NULL, start = NULL, end = NULL, shard = NULL,
  totalShards = NULL)
}
\arguments{
\item{projectId}{The Google Developers Console project ID or number which will be billed for this access}

\item{readGroupSetId}{The ID of the read group set from which to stream reads}

\item{referenceName}{The reference sequence name, for example `chr1`, `1`, or `chrX`}

\item{start}{The start position of the range on the reference, 0-based inclusive}

\item{end}{The end position of the range on the reference, 0-based exclusive}

\item{shard}{Restricts results to a shard containing approximately `1/totalShards` of the normal response payload for this query}

\item{totalShards}{Specifying `totalShards` causes a disjoint subset of the normal response payload to be returned for each query with a unique `shard` parameter specified}
}
\value{
StreamReadsRequest object
}
\description{
StreamReadsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The stream reads request.
}
\seealso{
Other StreamReadsRequest functions: \code{\link{reads.stream}}
}

