% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{StreamVariantsRequest}
\alias{StreamVariantsRequest}
\title{StreamVariantsRequest Object}
\usage{
StreamVariantsRequest(projectId = NULL, variantSetId = NULL,
  callSetIds = NULL, referenceName = NULL, start = NULL, end = NULL)
}
\arguments{
\item{projectId}{The Google Developers Console project ID or number which will be billed for this access}

\item{variantSetId}{The variant set ID from which to stream variants}

\item{callSetIds}{Only return variant calls which belong to call sets with these IDs}

\item{referenceName}{Required}

\item{start}{The beginning of the window (0-based, inclusive) for which overlapping variants should be returned}

\item{end}{The end of the window (0-based, exclusive) for which overlapping variants should be returned}
}
\value{
StreamVariantsRequest object
}
\description{
StreamVariantsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The stream variants request.
}
\seealso{
Other StreamVariantsRequest functions: \code{\link{variants.stream}}
}

