% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{UndeleteDatasetRequest}
\alias{UndeleteDatasetRequest}
\title{UndeleteDatasetRequest Object}
\usage{
UndeleteDatasetRequest()
}
\value{
UndeleteDatasetRequest object
}
\description{
UndeleteDatasetRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other UndeleteDatasetRequest functions: \code{\link{datasets.undelete}}
}

