% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{VariantAnnotation}
\alias{VariantAnnotation}
\title{VariantAnnotation Object}
\usage{
VariantAnnotation(type = NULL, effect = NULL, alternateBases = NULL,
  geneId = NULL, transcriptIds = NULL, conditions = NULL,
  clinicalSignificance = NULL)
}
\arguments{
\item{type}{Type has been adapted from ClinVar's list of variant types}

\item{effect}{Effect of the variant on the coding sequence}

\item{alternateBases}{The alternate allele for this variant}

\item{geneId}{Google annotation ID of the gene affected by this variant}

\item{transcriptIds}{Google annotation IDs of the transcripts affected by this variant}

\item{conditions}{The set of conditions associated with this variant}

\item{clinicalSignificance}{Describes the clinical significance of a variant}
}
\value{
VariantAnnotation object
}
\description{
VariantAnnotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

