% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{VariantCall.info}
\alias{VariantCall.info}
\title{VariantCall.info Object}
\usage{
VariantCall.info()
}
\value{
VariantCall.info object
}
\description{
VariantCall.info Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A map of additional variant call information. This must be of the form map (string key mapping to a list of string values).
}
\seealso{
Other VariantCall functions: \code{\link{VariantCall}}
}

