% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{VariantCall}
\alias{VariantCall}
\title{VariantCall Object}
\usage{
VariantCall(VariantCall.info = NULL, phaseset = NULL, info = NULL,
  callSetName = NULL, genotypeLikelihood = NULL, callSetId = NULL,
  genotype = NULL)
}
\arguments{
\item{VariantCall.info}{The \link{VariantCall.info} object or list of objects}

\item{phaseset}{If this field is present, this variant call's genotype ordering implies}

\item{info}{A map of additional variant call information}

\item{callSetName}{The name of the call set this variant call belongs to}

\item{genotypeLikelihood}{The genotype likelihoods for this variant call}

\item{callSetId}{The ID of the call set this variant call belongs to}

\item{genotype}{The genotype of this variant call}
}
\value{
VariantCall object
}
\description{
VariantCall Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A call represents the determination of genotype with respect to a particularvariant. It may include associated information such as quality and phasing.For example, a call might assign a probability of 0.32 to the occurrence ofa SNP named rs1234 in a call set with the name NA12345.
}
\seealso{
Other VariantCall functions: \code{\link{VariantCall.info}}
}
