% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_functions.R
\name{variants.import}
\alias{variants.import}
\title{Creates variant data by asynchronously importing the provided information.For the definitions of variant sets and other genomics resources, see[Fundamentals of GoogleGenomics](https://cloud.google.com/genomics/fundamentals-of-google-genomics)The variants for import will be merged with any existing variant thatmatches its reference sequence, start, end, reference bases, andalternative bases. If no such variant exists, a new one will be created.When variants are merged, the call information from the new variantis added to the existing variant, and Variant info fields are mergedas specified ininfoMergeConfig.As a special case, for single-sample VCF files, QUAL and FILTER fields willbe moved to the call level; these are sometimes interpreted in acall-specific context.Imported VCF headers are appended to the metadata already in a variant set.}
\usage{
variants.import(ImportVariantsRequest)
}
\arguments{
\item{ImportVariantsRequest}{The \link{ImportVariantsRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/devstorage.read_write
\item https://www.googleapis.com/auth/genomics
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/devstorage.read_write, https://www.googleapis.com/auth/genomics)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/genomics}{Google Documentation}

Other ImportVariantsRequest functions: \code{\link{ImportVariantsRequest.infoMergeConfig}},
  \code{\link{ImportVariantsRequest}}
}
