% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_functions.R
\name{variants.patch}
\alias{variants.patch}
\title{Updates a variant.For the definitions of variants and other genomics resources, see[Fundamentals of GoogleGenomics](https://cloud.google.com/genomics/fundamentals-of-google-genomics)This method supports patch semantics. Returns the modified variant withoutits calls.}
\usage{
variants.patch(Variant, variantId, updateMask = NULL)
}
\arguments{
\item{Variant}{The \link{Variant} object to pass to this method}

\item{variantId}{The ID of the variant to be updated}

\item{updateMask}{An optional mask specifying which fields to update}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/genomics
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/genomics)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/genomics}{Google Documentation}

Other Variant functions: \code{\link{Variant.info}},
  \code{\link{Variant}}, \code{\link{variants.create}}
}
