% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_functions.R
\name{variants.stream}
\alias{variants.stream}
\title{Returns a stream of all the variants matching the search request, ordered by reference name, position, and ID.}
\usage{
variants.stream(StreamVariantsRequest)
}
\arguments{
\item{StreamVariantsRequest}{The \link{StreamVariantsRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/genomics
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/genomics)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/genomics/}{Google Documentation}

Other StreamVariantsRequest functions: \code{\link{StreamVariantsRequest}}
}

