% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{VariantSet}
\alias{VariantSet}
\title{VariantSet Object}
\usage{
VariantSet(name = NULL, referenceSetId = NULL, metadata = NULL,
  referenceBounds = NULL, id = NULL, description = NULL,
  datasetId = NULL)
}
\arguments{
\item{name}{User-specified, mutable name}

\item{referenceSetId}{The reference set to which the variant set is mapped}

\item{metadata}{The metadata associated with this variant set}

\item{referenceBounds}{A list of all references used by the variants in a variant set}

\item{id}{The server-generated variant set ID, unique across all variant sets}

\item{description}{A textual description of this variant set}

\item{datasetId}{The dataset to which this variant set belongs}
}
\value{
VariantSet object
}
\description{
VariantSet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A variant set is a collection of call sets and variants. It contains summarystatistics of those contents. A variant set belongs to a dataset.For more genomics resource definitions, see [Fundamentals of GoogleGenomics](https://cloud.google.com/genomics/fundamentals-of-google-genomics)
}
\seealso{
Other VariantSet functions: \code{\link{variantsets.create}},
  \code{\link{variantsets.patch}}
}
