% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{VariantSetMetadata.info}
\alias{VariantSetMetadata.info}
\title{VariantSetMetadata.info Object}
\usage{
VariantSetMetadata.info()
}
\value{
VariantSetMetadata.info object
}
\description{
VariantSetMetadata.info Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Remaining structured metadata key-value pairs. This must be of the form map (string key mapping to a list of string values).
}
\seealso{
Other VariantSetMetadata functions: \code{\link{VariantSetMetadata}}
}

