% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{VariantSetMetadata}
\alias{VariantSetMetadata}
\title{VariantSetMetadata Object}
\usage{
VariantSetMetadata(VariantSetMetadata.info = NULL, key = NULL,
  value = NULL, id = NULL, type = NULL, number = NULL,
  description = NULL, info = NULL)
}
\arguments{
\item{VariantSetMetadata.info}{The \link{VariantSetMetadata.info} object or list of objects}

\item{key}{The top-level key}

\item{value}{The value field for simple metadata}

\item{id}{User-provided ID field, not enforced by this API}

\item{type}{The type of data}

\item{number}{The number of values that can be included in a field described by this metadata}

\item{description}{A textual description of this metadata}

\item{info}{Remaining structured metadata key-value pairs}
}
\value{
VariantSetMetadata object
}
\description{
VariantSetMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata describes a single piece of variant call metadata. These data include a top level key and either a single value string (value) or a list of key-value pairs (info.) Value and info are mutually exclusive.
}
\seealso{
Other VariantSetMetadata functions: \code{\link{VariantSetMetadata.info}}
}

