% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_functions.R
\name{variantsets.patch}
\alias{variantsets.patch}
\title{Updates a variant set using patch semantics. For the definitions of variant sets and other genomics resources, see [Fundamentals of Google Genomics](https://cloud.google.com/genomics/fundamentals-of-google-genomics)}
\usage{
variantsets.patch(VariantSet, variantSetId, updateMask = NULL)
}
\arguments{
\item{VariantSet}{The \link{VariantSet} object to pass to this method}

\item{variantSetId}{The ID of the variant to be updated (must already exist)}

\item{updateMask}{An optional mask specifying which fields to update}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/genomics
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/genomics)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/genomics/}{Google Documentation}

Other VariantSet functions: \code{\link{VariantSet}},
  \code{\link{variantsets.create}}
}

