% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{Annotation}
\alias{Annotation}
\title{Annotation Object}
\usage{
Annotation(Annotation.info = NULL, id = NULL, annotationSetId = NULL,
  name = NULL, referenceId = NULL, referenceName = NULL, start = NULL,
  end = NULL, reverseStrand = NULL, type = NULL, variant = NULL,
  transcript = NULL, info = NULL)
}
\arguments{
\item{Annotation.info}{The \link{Annotation.info} object or list of objects}

\item{id}{The server-generated annotation ID, unique across all annotations}

\item{annotationSetId}{The annotation set to which this annotation belongs}

\item{name}{The display name of this annotation}

\item{referenceId}{The ID of the Google Genomics reference associated with this range}

\item{referenceName}{The display name corresponding to the reference specified by `referenceId`, for example `chr1`, `1`, or `chrX`}

\item{start}{The start position of the range on the reference, 0-based inclusive}

\item{end}{The end position of the range on the reference, 0-based exclusive}

\item{reverseStrand}{Whether this range refers to the reverse strand, as opposed to the forward strand}

\item{type}{The data type for this annotation}

\item{variant}{A variant annotation, which describes the effect of a variant on the genome, the coding sequence, and/or higher level consequences at the organism level e}

\item{transcript}{A transcript value represents the assertion that a particular region of the reference genome may be transcribed as RNA}

\item{info}{A map of additional read alignment information}
}
\value{
Annotation object
}
\description{
Annotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An annotation describes a region of reference genome. The value of an annotation may be one of several canonical types, supplemented by arbitrary info tags. An annotation is not inherently associated with a specific sample or individual (though a client could choose to use annotations in this way). Example canonical annotation types are `GENE` and `VARIANT`.
}
\seealso{
Other Annotation functions: \code{\link{Annotation.info}},
  \code{\link{annotations.create}},
  \code{\link{annotations.update}}
}

