% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{AnnotationSet}
\alias{AnnotationSet}
\title{Genomics API Objects 
Stores, processes, explores and shares genomic data.}
\usage{
AnnotationSet(AnnotationSet.info = NULL, id = NULL, datasetId = NULL,
  referenceSetId = NULL, name = NULL, sourceUri = NULL, type = NULL,
  info = NULL)
}
\arguments{
\item{AnnotationSet.info}{The \link{AnnotationSet.info} object or list of objects}

\item{id}{The server-generated annotation set ID, unique across all annotation sets}

\item{datasetId}{The dataset to which this annotation set belongs}

\item{referenceSetId}{The ID of the reference set that defines the coordinate space for this set's annotations}

\item{name}{The display name for this annotation set}

\item{sourceUri}{The source URI describing the file from which this annotation set was generated, if any}

\item{type}{The type of annotations contained within this set}

\item{info}{A map of additional read alignment information}
}
\value{
AnnotationSet object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:19:40
filename: /Users/mark/dev/R/autoGoogleAPI/googlegenomicsv1.auto/R/genomics_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
AnnotationSet Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An annotation set is a logical grouping of annotations that share consistent type information and provenance. Examples of annotation sets include 'all genes from refseq', and 'all variant annotations from ClinVar'.
}
\seealso{
Other AnnotationSet functions: \code{\link{AnnotationSet.info}},
  \code{\link{annotationsets.create}},
  \code{\link{annotationsets.update}}
}

