% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{CigarUnit}
\alias{CigarUnit}
\title{CigarUnit Object}
\usage{
CigarUnit(operation = NULL, operationLength = NULL,
  referenceSequence = NULL)
}
\arguments{
\item{operation}{No description}

\item{operationLength}{The number of genomic bases that the operation runs for}

\item{referenceSequence}{`referenceSequence` is only used at mismatches (`SEQUENCE_MISMATCH`) and deletions (`DELETE`)}
}
\value{
CigarUnit object
}
\description{
CigarUnit Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A single CIGAR operation.
}

