% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ClinicalCondition}
\alias{ClinicalCondition}
\title{ClinicalCondition Object}
\usage{
ClinicalCondition(names = NULL, externalIds = NULL, conceptId = NULL,
  omimId = NULL)
}
\arguments{
\item{names}{A set of names for the condition}

\item{externalIds}{The set of external IDs for this condition}

\item{conceptId}{The MedGen concept id associated with this gene}

\item{omimId}{The OMIM id for this condition}
}
\value{
ClinicalCondition object
}
\description{
ClinicalCondition Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

