% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{Dataset}
\alias{Dataset}
\title{Dataset Object}
\usage{
Dataset(id = NULL, projectId = NULL, name = NULL, createTime = NULL)
}
\arguments{
\item{id}{The server-generated dataset ID, unique across all datasets}

\item{projectId}{The Google Developers Console project ID that this dataset belongs to}

\item{name}{The dataset name}

\item{createTime}{The time this dataset was created, in seconds from the epoch}
}
\value{
Dataset object
}
\description{
Dataset Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Dataset is a collection of genomic data. For more genomics resource definitions, see [Fundamentals of Google Genomics](https://cloud.google.com/genomics/fundamentals-of-google-genomics)
}
\seealso{
Other Dataset functions: \code{\link{datasets.create}},
  \code{\link{datasets.patch}}
}

