% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ExportReadGroupSetRequest}
\alias{ExportReadGroupSetRequest}
\title{ExportReadGroupSetRequest Object}
\usage{
ExportReadGroupSetRequest(exportUri = NULL, referenceNames = NULL,
  projectId = NULL)
}
\arguments{
\item{exportUri}{Required}

\item{referenceNames}{The reference names to export}

\item{projectId}{Required}
}
\value{
ExportReadGroupSetRequest object
}
\description{
ExportReadGroupSetRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The read group set export request.
}
\seealso{
Other ExportReadGroupSetRequest functions: \code{\link{readgroupsets.export}}
}
