% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ExportVariantSetRequest}
\alias{ExportVariantSetRequest}
\title{ExportVariantSetRequest Object}
\usage{
ExportVariantSetRequest(callSetIds = NULL, projectId = NULL,
  format = NULL, bigqueryDataset = NULL, bigqueryTable = NULL)
}
\arguments{
\item{callSetIds}{If provided, only variant call information from the specified call sets will be exported}

\item{projectId}{Required}

\item{format}{The format for the exported data}

\item{bigqueryDataset}{Required}

\item{bigqueryTable}{Required}
}
\value{
ExportVariantSetRequest object
}
\description{
ExportVariantSetRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The variant data export request.
}
\seealso{
Other ExportVariantSetRequest functions: \code{\link{variantsets.export}}
}

