% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_functions.R
\docType{package}
\name{genomics_googleAuthR}
\alias{genomics_googleAuthR}
\alias{genomics_googleAuthR-package}
\title{Genomics API
Upload, process, query, and search Genomics data in the cloud.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 19:53:38
filename: /Users/mark/dev/R/autoGoogleAPI/googlegenomicsv1.auto/R/genomics_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/bigquery
\item https://www.googleapis.com/auth/devstorage.read_write
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/genomics.readonly
\item https://www.googleapis.com/auth/genomics
}
}
