% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ImportVariantsRequest}
\alias{ImportVariantsRequest}
\title{ImportVariantsRequest Object}
\usage{
ImportVariantsRequest(ImportVariantsRequest.infoMergeConfig = NULL,
  variantSetId = NULL, sourceUris = NULL, format = NULL,
  normalizeReferenceNames = NULL, infoMergeConfig = NULL)
}
\arguments{
\item{ImportVariantsRequest.infoMergeConfig}{The \link{ImportVariantsRequest.infoMergeConfig} object or list of objects}

\item{variantSetId}{Required}

\item{sourceUris}{A list of URIs referencing variant files in Google Cloud Storage}

\item{format}{The format of the variant data being imported}

\item{normalizeReferenceNames}{Convert reference names to the canonical representation}

\item{infoMergeConfig}{A mapping between info field keys and the InfoMergeOperations to be performed on them}
}
\value{
ImportVariantsRequest object
}
\description{
ImportVariantsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The variant data import request.
}
\seealso{
Other ImportVariantsRequest functions: \code{\link{ImportVariantsRequest.infoMergeConfig}},
  \code{\link{variants.import}}
}

