% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_functions.R
\name{operations.list}
\alias{operations.list}
\title{Lists operations that match the specified filter in the request.}
\usage{
operations.list(name, pageToken = NULL, pageSize = NULL, filter = NULL)
}
\arguments{
\item{name}{The name of the operation collection}

\item{pageToken}{The standard list page token}

\item{pageSize}{The maximum number of results to return}

\item{filter}{A string for filtering Operations}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/genomics
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/genomics)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/genomics}{Google Documentation}
}
