% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{Range}
\alias{Range}
\title{Range Object}
\usage{
Range(start = NULL, end = NULL, referenceName = NULL)
}
\arguments{
\item{start}{The start position of the range on the reference, 0-based inclusive}

\item{end}{The end position of the range on the reference, 0-based exclusive}

\item{referenceName}{The reference sequence name, for example `chr1`,}
}
\value{
Range object
}
\description{
Range Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A 0-based half-open genomic coordinate range for search requests.
}
