% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ReadGroupSet}
\alias{ReadGroupSet}
\title{ReadGroupSet Object}
\usage{
ReadGroupSet(ReadGroupSet.info = NULL, info = NULL, id = NULL,
  datasetId = NULL, readGroups = NULL, filename = NULL, name = NULL,
  referenceSetId = NULL)
}
\arguments{
\item{ReadGroupSet.info}{The \link{ReadGroupSet.info} object or list of objects}

\item{info}{A map of additional read group set information}

\item{id}{The server-generated read group set ID, unique for all read group sets}

\item{datasetId}{The dataset to which this read group set belongs}

\item{readGroups}{The read groups in this set}

\item{filename}{The filename of the original source file for this read group set, if any}

\item{name}{The read group set name}

\item{referenceSetId}{The reference set to which the reads in this read group set are aligned}
}
\value{
ReadGroupSet object
}
\description{
ReadGroupSet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A read group set is a logical collection of read groups, which arecollections of reads produced by a sequencer. A read group set typicallymodels reads corresponding to one sample, sequenced one way, and aligned oneway.* A read group set belongs to one dataset.* A read group belongs to one read group set.* A read belongs to one read group.For more genomics resource definitions, see [Fundamentals of GoogleGenomics](https://cloud.google.com/genomics/fundamentals-of-google-genomics)
}
\seealso{
Other ReadGroupSet functions: \code{\link{ReadGroupSet.info}},
  \code{\link{readgroupsets.patch}}
}
