% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ReferenceSet}
\alias{ReferenceSet}
\title{ReferenceSet Object}
\usage{
ReferenceSet(description = NULL, sourceAccessions = NULL,
  ncbiTaxonId = NULL, sourceUri = NULL, referenceIds = NULL,
  md5checksum = NULL, assemblyId = NULL, id = NULL)
}
\arguments{
\item{description}{Free text description of this reference set}

\item{sourceAccessions}{All known corresponding accession IDs in INSDC (GenBank/ENA/DDBJ) ideally}

\item{ncbiTaxonId}{ID from http://www}

\item{sourceUri}{The URI from which the references were obtained}

\item{referenceIds}{The IDs of the reference objects that are part of this set}

\item{md5checksum}{Order-independent MD5 checksum which identifies this reference set}

\item{assemblyId}{Public id of this reference set, such as `GRCh37`}

\item{id}{The server-generated reference set ID, unique across all reference sets}
}
\value{
ReferenceSet object
}
\description{
ReferenceSet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A reference set is a set of references which typically comprise a referenceassembly for a species, such as `GRCh38` which is representativeof the human genome. A reference set defines a common coordinate space forcomparing reference-aligned experimental data. A reference set contains 1 ormore references.For more genomics resource definitions, see [Fundamentals of GoogleGenomics](https://cloud.google.com/genomics/fundamentals-of-google-genomics)
}
