% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{SearchAnnotationsRequest}
\alias{SearchAnnotationsRequest}
\title{SearchAnnotationsRequest Object}
\usage{
SearchAnnotationsRequest(annotationSetIds = NULL, referenceId = NULL,
  referenceName = NULL, start = NULL, end = NULL, pageToken = NULL,
  pageSize = NULL)
}
\arguments{
\item{annotationSetIds}{Required}

\item{referenceId}{The ID of the reference to query}

\item{referenceName}{The name of the reference to query, within the reference set associated with this query}

\item{start}{The start position of the range on the reference, 0-based inclusive}

\item{end}{The end position of the range on the reference, 0-based exclusive}

\item{pageToken}{The continuation token, which is used to page through large result sets}

\item{pageSize}{The maximum number of results to return in a single page}
}
\value{
SearchAnnotationsRequest object
}
\description{
SearchAnnotationsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other SearchAnnotationsRequest functions: \code{\link{annotations.search}}
}

