% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{Variant}
\alias{Variant}
\title{Variant Object}
\usage{
Variant(Variant.info = NULL, variantSetId = NULL, id = NULL,
  names = NULL, created = NULL, referenceName = NULL, start = NULL,
  end = NULL, referenceBases = NULL, alternateBases = NULL,
  quality = NULL, filter = NULL, info = NULL, calls = NULL)
}
\arguments{
\item{Variant.info}{The \link{Variant.info} object or list of objects}

\item{variantSetId}{The ID of the variant set this variant belongs to}

\item{id}{The server-generated variant ID, unique across all variants}

\item{names}{Names for the variant, for example a RefSNP ID}

\item{created}{The date this variant was created, in milliseconds from the epoch}

\item{referenceName}{The reference on which this variant occurs}

\item{start}{The position at which this variant occurs (0-based)}

\item{end}{The end position (0-based) of this variant}

\item{referenceBases}{The reference bases for this variant}

\item{alternateBases}{The bases that appear instead of the reference bases}

\item{quality}{A measure of how likely this variant is to be real}

\item{filter}{A list of filters (normally quality filters) this variant has failed}

\item{info}{A map of additional variant information}

\item{calls}{The variant calls for this particular variant}
}
\value{
Variant object
}
\description{
Variant Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A variant represents a change in DNA sequence relative to a reference sequence. For example, a variant could represent a SNP or an insertion. Variants belong to a variant set. For more genomics resource definitions, see [Fundamentals of Google Genomics](https://cloud.google.com/genomics/fundamentals-of-google-genomics) Each of the calls on a variant represent a determination of genotype with respect to that variant. For example, a call might assign probability of 0.32 to the occurrence of a SNP named rs1234 in a sample named NA12345. A call belongs to a call set, which contains related calls typically from one sample.
}
\seealso{
Other Variant functions: \code{\link{Variant.info}},
  \code{\link{variants.create}},
  \code{\link{variants.patch}}
}

