% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_functions.R
\name{variants.merge}
\alias{variants.merge}
\title{Merges the given variants with existing variants. For the definitions of variants and other genomics resources, see [Fundamentals of Google Genomics](https://cloud.google.com/genomics/fundamentals-of-google-genomics) Each variant will be merged with an existing variant that matches its reference sequence, start, end, reference bases, and alternative bases. If no such variant exists, a new one will be created. When variants are merged, the call information from the new variant is added to the existing variant. Variant info fields are merged as specified in the infoMergeConfig field of the MergeVariantsRequest. Please exercise caution when using this method! It is easy to introduce mistakes in existing variants and difficult to back out of them. For example, suppose you were trying to merge a new variant with an existing one and both variants contain calls that belong to callsets with the same callset ID. // Existing variant - irrelevant fields trimmed for clarity { 'variantSetId': '10473108253681171589', 'referenceName': '1', 'start': '10582', 'referenceBases': 'G', 'alternateBases': [ 'A' ], 'calls': [ { 'callSetId': '10473108253681171589-0', 'callSetName': 'CALLSET0', 'genotype': [ 0, 1 ], } ] } // New variant with conflicting call information { 'variantSetId': '10473108253681171589', 'referenceName': '1', 'start': '10582', 'referenceBases': 'G', 'alternateBases': [ 'A' ], 'calls': [ { 'callSetId': '10473108253681171589-0', 'callSetName': 'CALLSET0', 'genotype': [ 1, 1 ], } ] } The resulting merged variant would overwrite the existing calls with those from the new variant: { 'variantSetId': '10473108253681171589', 'referenceName': '1', 'start': '10582', 'referenceBases': 'G', 'alternateBases': [ 'A' ], 'calls': [ { 'callSetId': '10473108253681171589-0', 'callSetName': 'CALLSET0', 'genotype': [ 1, 1 ], } ] } This may be the desired outcome, but it is up to the user to determine if if that is indeed the case.}
\usage{
variants.merge(MergeVariantsRequest)
}
\arguments{
\item{MergeVariantsRequest}{The \link{MergeVariantsRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/genomics
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/genomics)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/genomics/}{Google Documentation}

Other MergeVariantsRequest functions: \code{\link{MergeVariantsRequest.infoMergeConfig}},
  \code{\link{MergeVariantsRequest}}
}

