% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{VariantSet}
\alias{VariantSet}
\title{VariantSet Object}
\usage{
VariantSet(datasetId = NULL, id = NULL, referenceSetId = NULL,
  referenceBounds = NULL, metadata = NULL, name = NULL,
  description = NULL)
}
\arguments{
\item{datasetId}{The dataset to which this variant set belongs}

\item{id}{The server-generated variant set ID, unique across all variant sets}

\item{referenceSetId}{The reference set to which the variant set is mapped}

\item{referenceBounds}{A list of all references used by the variants in a variant set with associated coordinate upper bounds for each one}

\item{metadata}{The metadata associated with this variant set}

\item{name}{User-specified, mutable name}

\item{description}{A textual description of this variant set}
}
\value{
VariantSet object
}
\description{
VariantSet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A variant set is a collection of call sets and variants. It contains summary statistics of those contents. A variant set belongs to a dataset. For more genomics resource definitions, see [Fundamentals of Google Genomics](https://cloud.google.com/genomics/fundamentals-of-google-genomics)
}
\seealso{
Other VariantSet functions: \code{\link{variantsets.create}},
  \code{\link{variantsets.patch}}
}

